renaming the Slices folder to backOffice in the store
renaming the Slices folder to backOffice in the store
