+++ src/Components/Backoffice/Slides/SlidesListBO.jsx
@@ -3,7 +3,7 @@
 import React, { useEffect } from 'react';
 import { useDispatch, useSelector } from 'react-redux';
 import { Link } from 'react-router-dom';
-import { getSlicesApi } from '../../../store/Slices/SlicesBackOffice';
+import { getSlicesApi } from '../../../store/backOffice/SlicesBackOffice';
 import SlideComponent from './SlideComponent';
 
 const SlideListBackOffice = () => {
